+++ lambdatimes/src/components/Content/Tab.js
@@ -1,22 +1,27 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 
-const Tab = props => {
-  /* Using your props, determine if the `tab` prop matches the `selectedTab` prop, 
+const Tab = (props) => {
+	/* Using your props, determine if the `tab` prop matches the `selectedTab` prop, 
       if they match, the className should be: 'tab active-tab', 
       if it is not it should just be 'tab'*/
-  return (
-    <div
-      className={''}
-      onClick={() => {
-        /* Replace this dummy click handler function with your selectTabHandler function from props 
-         you'll need to pass the `tab` in as an argument to this handler. */
-      }}
-    >
-      {props.tab.toUpperCase()}
-    </div>
-  );
+	return (
+		<div
+			id={props.tab}
+			className={props.tab === props.selectedTab ? 'tab active-tab' : 'tab'}
+			onClick={(event) => {
+				props.selectTabHandler(event);
+			}}
+		>
+			{props.tab}
+		</div>
+	);
 };
-
+Tab.propTypes = {
+	tab: PropTypes.string.isRequired,
+	selectedTab: PropTypes.string,
+	selectTabHandler: PropTypes.func
+};
 // Make sure you include PropTypes on your props.
 
 export default Tab;
